requirements.txt
pandas
python-dateutil
dateparser
streamlit
